+++ dikta-macos/Dikta/Models/AppConfig.swift
@@ -10,6 +10,7 @@ struct AppConfig: Codable {
     var llmModel: String
     var language: Language
     var customPrompt: String
+    var micDistance: MicDistance
     var muteSounds: Bool
     var muteNotifications: Bool
 
@@ -17,17 +18,20 @@ struct AppConfig: Codable {
         var toggle: HotkeyConfig
         var pushToTalk: HotkeyConfig
         var textToSpeech: HotkeyConfig
+        var languageToggle: HotkeyConfig
 
         enum CodingKeys: String, CodingKey {
             case toggle
             case pushToTalk = "push_to_talk"
             case textToSpeech = "text_to_speech"
+            case languageToggle = "language_toggle"
         }
 
-        init(toggle: HotkeyConfig, pushToTalk: HotkeyConfig, textToSpeech: HotkeyConfig = .defaultTextToSpeech) {
+        init(toggle: HotkeyConfig, pushToTalk: HotkeyConfig, textToSpeech: HotkeyConfig = .defaultTextToSpeech, languageToggle: HotkeyConfig = .defaultLanguageToggle) {
             self.toggle = toggle
             self.pushToTalk = pushToTalk
             self.textToSpeech = textToSpeech
+            self.languageToggle = languageToggle
         }
 
         init(from decoder: Decoder) throws {
@@ -35,6 +39,7 @@ struct AppConfig: Codable {
             toggle = try container.decode(HotkeyConfig.self, forKey: .toggle)
             pushToTalk = try container.decode(HotkeyConfig.self, forKey: .pushToTalk)
             textToSpeech = try container.decodeIfPresent(HotkeyConfig.self, forKey: .textToSpeech) ?? .defaultTextToSpeech
+            languageToggle = try container.decodeIfPresent(HotkeyConfig.self, forKey: .languageToggle) ?? .defaultLanguageToggle
         }
     }
 
@@ -47,6 +52,7 @@ struct AppConfig: Codable {
         case llmModel = "llm_model"
         case language
         case customPrompt = "custom_prompt"
+        case micDistance = "mic_distance"
         case muteSounds = "mute_sounds"
         case muteNotifications = "mute_notifications"
     }
@@ -67,6 +73,7 @@ struct AppConfig: Codable {
         llmModel: "gemma3",
         language: .english,
         customPrompt: defaultCustomPrompt,
+        micDistance: .normal,
         muteSounds: false,
         muteNotifications: false
     )
@@ -75,7 +82,7 @@ struct AppConfig: Codable {
     static let historyLimit = 5
 
     /// Memberwise initializer
-    init(version: Int, hotkeys: HotkeyConfigs, outputMode: OutputMode, history: [HistoryItem], whisperModel: String, llmModel: String, language: Language, customPrompt: String = defaultCustomPrompt, muteSounds: Bool = false, muteNotifications: Bool = false) {
+    init(version: Int, hotkeys: HotkeyConfigs, outputMode: OutputMode, history: [HistoryItem], whisperModel: String, llmModel: String, language: Language, customPrompt: String = defaultCustomPrompt, micDistance: MicDistance = .normal, muteSounds: Bool = false, muteNotifications: Bool = false) {
         self.version = version
         self.hotkeys = hotkeys
         self.outputMode = outputMode
@@ -84,6 +91,7 @@ struct AppConfig: Codable {
         self.llmModel = llmModel
         self.language = language
         self.customPrompt = customPrompt
+        self.micDistance = micDistance
         self.muteSounds = muteSounds
         self.muteNotifications = muteNotifications
     }
@@ -108,6 +116,7 @@ struct AppConfig: Codable {
         llmModel = try container.decode(String.self, forKey: .llmModel)
         language = try container.decodeIfPresent(Language.self, forKey: .language) ?? .english
         customPrompt = try container.decodeIfPresent(String.self, forKey: .customPrompt) ?? Self.defaultCustomPrompt
+        micDistance = try container.decodeIfPresent(MicDistance.self, forKey: .micDistance) ?? .normal
         muteSounds = try container.decodeIfPresent(Bool.self, forKey: .muteSounds) ?? false
         muteNotifications = try container.decodeIfPresent(Bool.self, forKey: .muteNotifications) ?? false
     }